# api/openapi.yaml
openapi: 3.0.3
info: {title: Netlab API, version: 1.0.0}
servers: [{url: /api/v1}]
security: [{bearer: []}]
paths:
...
  /identity/password:
    put:
      operationId: changePassword
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ChangePassword'}}}}
      responses: {'204': {description: Password changed; sign in again}}
  /principals/{id}/grants:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get:
      operationId: principalGrants
      responses:
        '200': {description: Account grants, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/ScopeGrant'}}}}}
...
  /principals:
    get:
      operationId: listPrincipals
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Search'}, {name: kind, in: query, schema: {type: string, enum: [user, token]}}]
      responses:
        '200': {description: Accounts, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Principal'}}}}}
    post:
      operationId: createUser
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateUser'}}}}
      responses:
        '201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/Principal'}}}}
  /principals/{id}:
    parameters: [{$ref: '#/components/parameters/Id'}]
    put:
      operationId: updateUser
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/UpdateUser'}}}}
      responses: {'204': {description: Updated}}
  /service-tokens:
    post:
      operationId: createServiceToken
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/CreateServiceToken'}}}}
      responses:
        '201': {description: Token returned once, content: {application/json: {schema: {$ref: '#/components/schemas/IssuedServiceToken'}}}}
...
  /environments/{id}/assets/{assetId}/logs:
    parameters:
      - {$ref: '#/components/parameters/Id'}
      - {name: assetId, in: path, required: true, schema: {type: string}}
    get:
      operationId: assetLogs
      description: Container stdout/stderr as output events. Initial tail is bounded to 1 MiB per stream; follow continues until disconnected or access is revoked.
      parameters:
        - {name: tail, in: query, schema: {type: integer, minimum: 0, maximum: 10000, default: 200}}
        - {name: stream, in: query, schema: {type: string, enum: [all, stdout, stderr], default: all}}
        - {name: follow, in: query, schema: {type: boolean, default: false}}
      responses:
        '200': {description: Log stream, content: {text/event-stream: {schema: {$ref: '#/components/schemas/LogChunk'}}}}
        '409': {description: Asset is not a container}
  /operations:
    get:
      operationId: listOperations
      parameters: [{name: environmentId, in: query, schema: {type: string}}, {$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}]
      responses:
        '200': {description: Operations, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Operation'}}}}}
...
  /templates:
    get:
      operationId: listTemplates
      parameters:
        - {$ref: '#/components/parameters/Cursor'}
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Search'}
        - {name: kind, in: query, schema: {$ref: '#/components/schemas/TemplateKind'}}
        - {name: ids, in: query, style: form, explode: false, schema: {type: array, items: {type: string}}}
      responses:
        '200': {description: Templates, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Template'}}}}}
    post:
      operationId: createTemplate
      requestBody:
        required: true
        description: Send the template JSON part first. Source names the main file, relative to the uploaded files; preserve relative filenames for OVF and split VMDK.
        content:
          application/json: {schema: {$ref: '#/components/schemas/TemplateImport'}}
          multipart/form-data:
            schema:
              type: object
              required: [template, files]
              properties:
                template: {$ref: '#/components/schemas/TemplateImport'}
                files: {type: array, items: {type: string, format: binary}}
            encoding:
              template: {contentType: application/json}
      responses:
        '201':
          description: Created
          headers:
            Operation-Location: {description: Template import operation URL, schema: {type: string}}
          content: {application/json: {schema: {$ref: '#/components/schemas/Template'}}}
...
  /nodes:
    get:
      operationId: listNodes
      parameters: [{$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Search'}]
      responses:
        '200': {description: Nodes, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Node'}}}}}
    post:
      operationId: registerNode
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/NodeRegistration'}}}}
      responses:
        '201': {description: Registered, content: {application/json: {schema: {$ref: '#/components/schemas/Node'}}}}
...
components:
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string}}
    Cursor: {name: cursor, in: query, schema: {type: string}}
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 500, default: 100}}
    Search: {name: search, in: query, schema: {type: string}}
...
  securitySchemes:
    bearer: {type: http, scheme: bearer}
  schemas:
...
    Permission:
      type: string
      enum: [read, operate, compose, manage, session, file, observe, network, access]
    ScopeGrant:
      type: object
      required: [scopeKind, scopeId, permissions]
      properties:
        scopeKind: {type: string, enum: [project, environment, asset]}
        scopeId: {type: string}
        permissions: {type: array, items: {$ref: '#/components/schemas/Permission'}}
...
    Principal:
      type: object
      required: [id, name, kind, administrator, disabled, createdAt]
      properties:
        id: {type: string}
        name: {type: string}
        kind: {type: string, enum: [user, token]}
        administrator: {type: boolean}
        disabled: {type: boolean}
        createdAt: {type: string, format: date-time}
        expiresAt: {type: string, format: date-time}
    CreateUser:
      type: object
      required: [name, password]
      properties:
        name: {type: string, minLength: 1, maxLength: 100}
        password: {type: string, minLength: 8, maxLength: 72, format: password}
        administrator: {type: boolean, x-go-type-skip-optional-pointer: true}
        grants: {type: array, items: {$ref: '#/components/schemas/ScopeGrant'}, x-go-type-skip-optional-pointer: true}
    UpdateUser:
      type: object
      required: [name, disabled]
      properties:
        name: {type: string, minLength: 1, maxLength: 100}
        password: {type: string, minLength: 8, maxLength: 72, format: password}
        disabled: {type: boolean}
        administrator: {type: boolean}
        grants: {type: array, items: {$ref: '#/components/schemas/ScopeGrant'}}
    ChangePassword:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: {type: string, format: password}
        newPassword: {type: string, minLength: 8, maxLength: 72, format: password}
    CreateServiceToken:
      type: object
      required: [name, grants]
      properties:
        name: {type: string, minLength: 1, maxLength: 100}
        grants: {type: array, minItems: 1, items: {$ref: '#/components/schemas/ScopeGrant'}}
        expiresAt: {type: string, format: date-time}
    IssuedServiceToken:
      type: object
      required: [principal, token]
      properties:
        principal: {$ref: '#/components/schemas/Principal'}
        token: {type: string}
...
    Resources:
      type: object
      required: [cpu, memoryMiB, diskGiB]
      properties:
        cpu: {type: integer, minimum: 1}
        memoryMiB: {type: integer, format: int64, minimum: 64}
        diskGiB: {type: integer, format: int64, minimum: 1}
    Hardware:
      type: object
      required: [firmware, machine, diskBus, nicModel]
      properties:
        firmware: {type: string, enum: [bios, uefi]}
        machine: {type: string}
        diskBus: {type: string, enum: [ide, sata, scsi, virtio]}
        diskController: {type: string}
        nicModel: {type: string, enum: [virtio, e1000, e1000e, rtl8139, vmxnet3]}
        cpuModel: {type: string}
        cpuTopology: {$ref: '#/components/schemas/CpuTopology'}
        numaNodes: {type: integer, minimum: 1, description: Guest NUMA nodes; vCPUs and memory are divided evenly. Omitted keeps one NUMA node.}
        numaPlacement: {type: string, enum: [preferred, strict], description: Native automatic host CPU and memory placement. Omitted lets the host scheduler choose.}
        secureBoot: {type: boolean}
        tpm: {type: boolean}
        guestAgent: {type: boolean}
        firmwareCode: {type: string}
        firmwareVars: {type: string}
    CpuTopology:
      type: object
      required: [sockets, threads]
      properties:
        sockets: {type: integer, minimum: 1}
        threads: {type: integer, minimum: 1}
    Volume:
      type: object
      required: [id, mountPath, sizeGiB]
      properties:
        id: {type: string}
        mountPath: {type: string}
        sizeGiB: {type: integer, format: int64}
        persistentVolumeId: {type: string, description: Registered persistent volume. Its size comes from the volume catalogue.}
...
    NodeVolume:
      type: object
      required: [id, kind, sizeGiB, storage]
      properties:
        id: {type: string}
        kind: {$ref: '#/components/schemas/TemplateKind'}
        sizeGiB: {type: integer, format: int64}
        storage: {$ref: '#/components/schemas/StorageInfo'}
...
    TemplateDisk:
      type: object
      required: [id, sizeGiB, bus, bootOrder]
      properties:
        id: {type: string}
        sizeGiB: {type: integer, format: int64, minimum: 1}
        bus: {type: string, enum: [ide, sata, scsi, virtio]}
        bootOrder: {type: integer, minimum: 1}
        controllerModel: {type: string}
        controllerIndex: {type: integer, minimum: 0}
        controllerUnit: {type: integer, minimum: 0}
    RegistryCredentials:
      type: object
      properties:
        username: {type: string}
        password: {type: string, format: password, writeOnly: true}
        plainHttp: {type: boolean, default: false, description: Explicitly connect to this registry over HTTP}
    TemplateImport:
      allOf:
        - {$ref: '#/components/schemas/Template'}
        - type: object
          properties:
            registry: {$ref: '#/components/schemas/RegistryCredentials'}
    TemplateFormat:
      type: string
      enum: [oci, docker, qcow2, raw, vmdk, ova, ovf, iso]
    TemplateInitialization:
      type: string
      enum: [none, cloud-init, cloudbase-init]
      default: none
    TemplateState:
      type: string
      enum: [importing, ready, failed, deleting]
      x-enum-varnames: [TemplateStateImporting, TemplateStateReady, TemplateStateFailed, TemplateStateDeleting]
    Template:
      type: object
      required: [id, name, kind, os, version, source, resources]
      properties:
        id: {type: string}
        name: {type: string}
        kind: {$ref: '#/components/schemas/TemplateKind'}
        os: {type: string}
        version: {type: integer}
        source: {type: string}
        format: {$ref: '#/components/schemas/TemplateFormat'}
        artifactNodeId: {type: string, readOnly: true}
        operationId: {type: string, readOnly: true}
        resources: {$ref: '#/components/schemas/Resources'}
        hardware: {$ref: '#/components/schemas/Hardware'}
        disks: {type: array, items: {$ref: '#/components/schemas/TemplateDisk'}}
        media: {type: array, items: {$ref: '#/components/schemas/TemplateMedia'}}
        stateFiles: {type: array, items: {type: string}, readOnly: true}
        nicModels: {type: array, items: {type: string}}
        initialization: {$ref: '#/components/schemas/TemplateInitialization'}
        volumes: {type: array, items: {$ref: '#/components/schemas/Volume'}}
        state: {$ref: '#/components/schemas/TemplateState'}
        error: {type: string}
    TemplateMedia:
      type: object
      required: [id, source]
      properties:
        id: {type: string}
        source: {type: string}
...
    Interface:
      type: object
      required: [id, networkId, mac, address, primary]
      properties:
        id: {type: string}
        networkId: {type: string}
        mac: {type: string}
        address: {type: string}
        primary: {type: boolean}
    Asset:
      type: object
      required: [id, name, templateId, resources, interfaces]
      properties:
        id: {type: string}
        name: {type: string}
        templateId: {type: string}
        storagePoolId: {type: string, description: Omitted selects storage automatically; an existing asset keeps its storage until migrated.}
        resources: {$ref: '#/components/schemas/Resources'}
        interfaces: {type: array, items: {$ref: '#/components/schemas/Interface'}}
        volumes: {type: array, items: {$ref: '#/components/schemas/Volume'}}
        parameters: {type: object, additionalProperties: {type: string}}
        restartPolicy: {$ref: '#/components/schemas/RestartPolicy'}
        guest: {$ref: '#/components/schemas/GuestSettings'}
        media: {type: array, items: {type: string}, description: Attached template media IDs. Omitted attaches all; an empty array ejects all.}
        pciBinding: {$ref: '#/components/schemas/PciBinding'}
...
    PciGroup:
      type: object
      required: [id, name, available, devices]
      properties:
        id: {type: string}
        name: {type: string}
        available: {type: boolean}
        devices: {type: array, items: {type: string}}
    RestartPolicy:
      type: string
      enum: [never, on-failure, always]
      default: never
      description: Restart unexpected exits only. Explicit stop remains stopped, including across node restarts.
    LogChunk:
      type: object
      required: [stream, data]
      properties:
        stream: {type: string, enum: [stdout, stderr]}
        data: {type: string}
...
    TemplateKind:
      type: string
      enum: [container, vm]
...
    Operation:
      type: object
      required: [id, kind, state, phase, completed, total, retryable, createdAt, updatedAt]
      properties:
        id: {type: string}
        environmentId: {type: string}
        kind: {type: string}
        state: {type: string, enum: [queued, running, succeeded, failed, partially_applied]}
        phase: {type: string}
        completed: {type: integer}
        total: {type: integer}
        retryable: {type: boolean}
        error: {type: string}
        results: {type: array, items: {$ref: '#/components/schemas/ExecutionResult'}}
        createdAt: {type: string, format: date-time}
        updatedAt: {type: string, format: date-time}
    Node:
      type: object
      required: [id, name, endpoint, capacity, reserved, capabilities, slots, observedAt]
      properties:
        id: {type: string}
        name: {type: string}
        endpoint: {type: string}
        capacity: {$ref: '#/components/schemas/Resources'}
        reserved: {$ref: '#/components/schemas/Resources'}
        capabilities: {type: array, items: {type: string}}
        vmHardware: {$ref: '#/components/schemas/VmHardware'}
        storageDevice: {type: string, description: Dedicated Ceph disk selected during node installation}
        slots: {type: integer}
        observedAt: {type: string, format: date-time}
        state: {type: string}
        override: {$ref: '#/components/schemas/Resources'}
    ResolvedInterface:
      type: object
      required: [id, networkId, mac, address, prefix, primary, portName, mtu]
      properties:
        id: {type: string}
        networkId: {type: string}
        mac: {type: string}
        address: {type: string}
        prefix: {type: integer}
        primary: {type: boolean}
        portName: {type: string}
        mtu: {type: integer}
        gateway: {type: string}
        dns: {type: array, items: {type: string}}
    AssetExecution:
      type: object
      required: [asset, template, instanceId, interfaces]
      properties:
        asset: {$ref: '#/components/schemas/Asset'}
        template: {$ref: '#/components/schemas/Template'}
        instanceId: {type: string}
        previousInstanceId: {type: string}
        dataSetId: {type: string, x-go-type-skip-optional-pointer: true, description: Actual writable data set; retained across ordinary rebuilds}
        volumeSources: {type: object, additionalProperties: {$ref: '#/components/schemas/NodeVolume'}, x-go-type-skip-optional-pointer: true}
        pciDevices: {type: array, items: {type: string}, x-go-type-skip-optional-pointer: true}
        storagePath: {type: string, description: Node-local managed storage selected by the scheduler}
        storagePoolId: {type: string, description: Actual selected pool; omitted for node default storage}
        storageFilesystem: {type: string, description: Physical filesystem reservation identity}
        rbd: {$ref: '#/components/schemas/RbdStorage'}
        interfaces: {type: array, items: {$ref: '#/components/schemas/ResolvedInterface'}}
...
    ExecutionResult:
      type: object
      required: [assetId, instanceId, state, observedAt]
      properties:
        environmentId: {type: string}
        nodeId: {type: string}
        assetId: {type: string}
        instanceId: {type: string}
        state: {type: string}
        error: {type: string}
        execution: {$ref: '#/components/schemas/AssetExecution'}
        recovery: {$ref: '#/components/schemas/RecoveryCapture'}
        observedAt: {type: string, format: date-time}
    RecoveryConsistency:
      type: string
      enum: [crash, filesystem, application]
    RecoveryCapture:
      type: object
      required: [sizeBytes, memory, consistency]
      properties:
        sizeBytes: {type: integer, format: int64}
        memory: {type: boolean}
        consistency: {$ref: '#/components/schemas/RecoveryConsistency'}
...
    RbdStorage:
      type: object
      required: [pool, imagePrefix, monitors, user, fsid, secretId]
      properties:
        pool: {type: string}
        imagePrefix: {type: string}
        monitors: {type: array, items: {type: string}}
        user: {type: string}
        fsid: {type: string}
        secretId: {type: string}
...
    VmHardware:
      type: object
      required: [machines, cpuModes, cpuModels, nicModels, diskControllers]
      properties:
        machines: {type: array, items: {$ref: '#/components/schemas/VmMachine'}}
        cpuModes: {type: array, items: {type: string}}
        cpuModels: {type: array, items: {type: string}}
        nicModels: {type: array, items: {type: string}}
        diskControllers: {type: array, items: {type: string}}
        numaPlacement: {type: boolean, description: Host supports libvirt automatic placement through numad.}
        pciGroups: {type: array, items: {$ref: '#/components/schemas/PciGroup'}, x-go-type-skip-optional-pointer: true}
    VmMachine:
      type: object
      required: [name, aliases, maxVcpus, firmware, secureBoot, tpm2, diskBuses, firmwareFiles]
      properties:
        name: {type: string}
        aliases: {type: array, items: {type: string}}
        maxVcpus: {type: integer}
        firmware: {type: array, items: {type: string}}
        secureBoot: {type: boolean}
        tpm2: {type: boolean}
        diskBuses: {type: array, items: {type: string}}
        firmwareFiles: {type: array, items: {type: string}}
...
    NodeRegistration:
      type: object
      required: [name, endpoint]
      properties:
        name: {type: string}
        endpoint: {type: string}
...